move store.go to store pkg
