+++ x/pokt/client/cmd/deleteChainHistory.go
@@ -18,7 +18,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/lostak/pokt/keeper"
+	"github.com/lostak/pokt/store"
 	"github.com/spf13/cobra"
 )
 
@@ -30,7 +30,7 @@ var deleteChainHistoryCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("deleteChainHistory called")
 
-		portfolio, err := keeper.GetPortfolio()
+		portfolio, err := store.GetPortfolio()
 		if err != nil {
 			fmt.Println(err.Error())
 			return
@@ -42,7 +42,7 @@ var deleteChainHistoryCmd = &cobra.Command{
 			return
 		}
 
-		if err := keeper.SetPortfolio(portfolio); err != nil {
+		if err := store.SetPortfolio(portfolio); err != nil {
 			fmt.Println(err.Error())
 			return
 		}
