+++ x/pokt/client/cmd/print.go
@@ -18,7 +18,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/lostak/pokt/keeper"
+	"github.com/lostak/pokt/store"
 	"github.com/spf13/cobra"
 )
 
@@ -29,7 +29,7 @@ var printCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Println("print called")
 
-		portfolio, err := keeper.GetPortfolio()
+		portfolio, err := store.GetPortfolio()
 		if err != nil {
 			fmt.Println(err.Error())
 			return
