+++ x/pokt/store/store.go
@@ -1,4 +1,4 @@
-package keeper
+package store
 
 import (
 	"errors"
